refactor: relocate finder entry points into their submodules
Move the promoted `pathfinder` and `neighborhood_finder` functions out of
`TCT/TCT.py` into `TCT/TCT_pathfinder.py` and `TCT/TCT_neighborhood_finder.py`,
following the repo convention of keeping finder logic in its submodule and
importing/re-exporting the public names from `TCT`. The shared infrastructure
stays in `TCT.py`: the result/resource dataclasses (`FinderResult`,
`ResolvedNode`, `TranslatorResources`), `get_translator_resources` /
`clear_translator_resource_cache`, and the node/category/resource helpers.

`neighborhood_finder` is defined as `_neighborhood_finder` in its submodule
(to avoid colliding with the existing low-level
`TCT_neighborhood_finder.neighborhood_finder` used by the network annotator
and notebooks) and re-exported under its public name from `TCT.py`, so no
legacy caller changes. The submodule re-exports sit at the end of `TCT.py`
(after the helpers are defined) to avoid a circular import. Retarget the
finder-test monkeypatches at the submodules; helper tests are unchanged.
