+++ TCT/TCT_neighborhood_finder.py
@@ -1,8 +1,23 @@
 from collections import Counter
 
+import pandas as pd
 from .TCT import sele_predicates_API, parse_KG, rank_by_primary_infores
 from .TCT_pathfinder import generate_score_results, build_query_graph
 from .translator_query import format_query_json
+
+from typing import Any, Optional, Union
+
+from . import translator_query
+from .TCT import (
+    CategoryList,
+    FinderResult,
+    NodeInput,
+    TranslatorResources,
+    _build_finder_result,
+    _get_resources,
+    _normalize_categories,
+    _resolve_nodes,
+)
 
 def parse_results_for_neighborhood_finder(start_node_id:str, results:dict,
         start_node_categories=None, end_node_categories=None,
@@ -371,3 +386,123 @@ def neighborhood_finder_multiple_inputs(input_nodes:list[str], node2_categories:
     return result, parsed_results
 
 
+def _neighborhood_finder(
+    node: Union[NodeInput, list[NodeInput]],
+    neighbor_categories: CategoryList,
+    *,
+    node_categories: Optional[CategoryList] = None,
+    api_names: Optional[dict[str, str]] = None,
+    meta_kg: Optional[pd.DataFrame] = None,
+    api_predicates: Optional[dict[str, list[str]]] = None,
+    resources: Optional[TranslatorResources] = None,
+    predicates_subset: Optional[list[str]] = None,
+    attribute_constraints: Optional[list[dict[str, Any]]] = None,
+    name_resolver_kwargs: Optional[dict[str, Any]] = None,
+    node_normalizer_kwargs: Optional[dict[str, Any]] = None,
+) -> FinderResult:
+    """
+    Find one-hop neighbors for one or more biomedical concepts.
+
+    Parameters
+    ----------
+    node : str or list[str]
+        Source node or nodes. Each value may be a CURIE or human-readable
+        string. Human-readable strings are resolved with Name Resolver and then
+        normalized with Node Normalizer.
+    neighbor_categories : list[str]
+        Desired neighbor categories. Values may be short names like ``"Drug"``
+        or full Biolink names like ``"biolink:Drug"``.
+    node_categories : list[str], optional
+        Category override for source nodes. If omitted, categories are inferred
+        from the first normalized source node.
+    resources : TranslatorResources, optional
+        Preloaded Translator resources. If omitted, the module-level singleton
+        is loaded on first use and reused.
+    api_names, meta_kg, api_predicates : optional
+        Advanced partial overrides for the Translator resources used by the
+        neighborhood implementation.
+    predicates_subset : list[str], optional
+        Optional predicate filter applied after MetaKG predicate selection.
+    attribute_constraints : list[dict], optional
+        TRAPI attribute constraints passed through to query construction.
+    name_resolver_kwargs : dict, optional
+        Extra keyword arguments for ``name_resolver.lookup``.
+    node_normalizer_kwargs : dict, optional
+        Extra keyword arguments for ``node_normalizer.get_normalized_nodes``.
+
+    Returns
+    -------
+    FinderResult
+        Convenience wrapper containing resolved input nodes, parsed neighborhood
+        knowledge graph, results, auxiliary graphs, and raw TRAPI-style output.
+
+    Examples
+    --------
+    >>> from TCT import neighborhood_finder
+    >>> result = neighborhood_finder("asthma", ["SmallMolecule", "Drug"])
+    >>> result.knowledge_graph["nodes"]
+    {...}
+    """
+    resolved_nodes = _resolve_nodes(
+        node,
+        name_resolver_kwargs=name_resolver_kwargs,
+        node_normalizer_kwargs=node_normalizer_kwargs,
+    )
+    source_categories = (
+        _normalize_categories(node_categories) or resolved_nodes[0].categories
+    )
+    neighbor_categories = _normalize_categories(neighbor_categories) or []
+    resolved_resources = _get_resources(
+        resources=resources,
+        api_names=api_names,
+        meta_kg=meta_kg,
+        api_predicates=api_predicates,
+    )
+
+    predicates, apis, _ = sele_predicates_API(
+        source_categories,
+        neighbor_categories,
+        resolved_resources.meta_kg,
+        resolved_resources.api_names,
+    )
+    if predicates_subset is not None:
+        predicates = list(set(predicates).intersection(predicates_subset))
+    if len(predicates) == 0:
+        predicates = ["biolink:related_to"]
+
+    input_curies = [resolved_node.curie for resolved_node in resolved_nodes]
+    query = translator_query.format_query_json(
+        subject_ids=input_curies,
+        object_ids=None,
+        subject_categories=None,
+        object_categories=neighbor_categories,
+        predicates=predicates,
+        attribute_constraints=attribute_constraints,
+    )
+    raw_edges = translator_query.parallel_api_query(
+        query_json=query,
+        select_APIs=apis,
+        APInames=resolved_resources.api_names,
+        API_predicates=resolved_resources.api_predicates,
+        max_workers=max(1, len(apis)),
+    )
+    if isinstance(node, str):
+        raw_output = parse_results_for_neighborhood_finder(
+            input_curies[0],
+            raw_edges,
+            source_categories,
+            neighbor_categories,
+        )
+        result_nodes = {"node": resolved_nodes[0]}
+    else:
+        raw_output = parse_results_for_neighborhood_finder_multiple_inputs(
+            input_curies,
+            raw_edges,
+            source_categories,
+            neighbor_categories,
+        )
+        result_nodes = {
+            f"node_{index}": resolved_node
+            for index, resolved_node in enumerate(resolved_nodes)
+        }
+    return _build_finder_result(raw_output, resolved_nodes=result_nodes)
